fix(context): address review — top_files paths + layering cleanup
Codex review (P2): the lean prior-session loader projected `<string>out AS
file` (a `file:...` record id), so the pack's "Files:" line rendered record
ids instead of paths. The old slow loader projected `file.path`. The hook
never rendered top_files so it was invisible; the pack does. Project
`out.path AS file` and pin it with a top_files regression assertion.

Simplify cleanup (3 agents converged):
- Extract context/file-evidence-types.ts as the bottom layer (row/result types
  + BuildFileContextInput). The pure rank module now imports DOWN instead of
  up from the DB-query module; file-evidence.ts re-exports the types so both
  adapters keep one import site.
- Drop the dead `export type {...}` re-export in file-context-hook.ts (no
  consumer; telemetry.ts imports the types from file-evidence directly).
- Replace the hook's duplicate `clipText` with the exported `clip`.

Kept (documented, user-approved): hands_free_ms->null and pr-title-first
prior-session title — both are the single-loader consolidation, not bugs;
restoring them would re-fork the loader / re-add the slow phase_span query.

Gates: bun test (4731 pass / 0 fail) + bun run typecheck (clean).
